#8051 - Generic dto to handle line entities
