+++ sormas-ui/src/main/java/de/symeda/sormas/ui/contact/ContactController.java
@@ -22,7 +22,6 @@ import java.util.function.Consumer;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
-import de.symeda.sormas.ui.utils.NotificationHelper;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,14 +29,11 @@ import org.slf4j.LoggerFactory;
 import com.vaadin.navigator.Navigator;
 import com.vaadin.server.Page;
 import com.vaadin.server.Sizeable.Unit;
-import com.vaadin.ui.Alignment;
-import com.vaadin.ui.Button;
 import com.vaadin.ui.Label;
 import com.vaadin.ui.Notification;
 import com.vaadin.ui.Notification.Type;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.Window;
-import com.vaadin.ui.themes.ValoTheme;
 
 import de.symeda.sormas.api.Disease;
 import de.symeda.sormas.api.DiseaseHelper;
@@ -82,13 +78,14 @@ import de.symeda.sormas.ui.contact.components.linelisting.layout.LineListingLayo
 import de.symeda.sormas.ui.epidata.ContactEpiDataView;
 import de.symeda.sormas.ui.epidata.EpiDataForm;
 import de.symeda.sormas.ui.utils.AbstractView;
-import de.symeda.sormas.ui.utils.ButtonHelper;
 import de.symeda.sormas.ui.utils.CommitDiscardWrapperComponent;
 import de.symeda.sormas.ui.utils.CoreEntityArchiveMessages;
 import de.symeda.sormas.ui.utils.CssStyles;
+import de.symeda.sormas.ui.utils.NotificationHelper;
 import de.symeda.sormas.ui.utils.VaadinUiUtil;
 import de.symeda.sormas.ui.utils.ViewMode;
 import de.symeda.sormas.ui.utils.components.automaticdeletion.AutomaticDeletionLabel;
+import de.symeda.sormas.ui.utils.components.linelisting.model.LineDto;
 import de.symeda.sormas.ui.utils.components.page.title.RowLayout;
 import de.symeda.sormas.ui.utils.components.page.title.TitleLayout;
 
@@ -134,7 +131,7 @@ public class ContactController {
 		UI.getCurrent().addWindow(window);
 	}
 
-	private void saveContactsFromLineListing(LineListingLayout lineListingForm, List<LineListingLayout.ContactLineDto> contacts) {
+	private void saveContactsFromLineListing(LineListingLayout lineListingForm, List<LineDto<ContactDto>> contacts) {
 		try {
 			lineListingForm.validate();
 		} catch (ValidationRuntimeException e) {
@@ -142,8 +139,8 @@ public class ContactController {
 			return;
 		}
 
-		for (LineListingLayout.ContactLineDto contactLineDto : contacts) {
-			ContactDto newContact = contactLineDto.getContact();
+		for (LineDto<ContactDto> contactLineDto : contacts) {
+			ContactDto newContact = contactLineDto.getEntity();
 			if (UserProvider.getCurrent() != null) {
 				newContact.setReportingUser(UserProvider.getCurrent().getUserReference());
 			}
@@ -193,7 +190,7 @@ public class ContactController {
 		UI.getCurrent().addWindow(window);
 	}
 
-	private void saveContactsFromEventParticipantsLineListing(LineListingLayout lineListingForm, List<LineListingLayout.ContactLineDto> contacts) {
+	private void saveContactsFromEventParticipantsLineListing(LineListingLayout lineListingForm, List<LineDto<ContactDto>> contacts) {
 		try {
 			lineListingForm.validate();
 		} catch (ValidationRuntimeException e) {
@@ -201,8 +198,8 @@ public class ContactController {
 			return;
 		}
 
-		for (LineListingLayout.ContactLineDto contactLineDto : contacts) {
-			ContactDto newContact = contactLineDto.getContact();
+		for (LineDto<ContactDto> contactLineDto : contacts) {
+			ContactDto newContact = contactLineDto.getEntity();
 			if (UserProvider.getCurrent() != null) {
 				newContact.setReportingUser(UserProvider.getCurrent().getUserReference());
 			}
